# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Phylax API
  version: 0.1.0
  description: Demo-first risk, payment, rebalance, and agent service API.
paths:
...
  /api/agent/query:
    post:
      summary: Agent intent endpoint
      responses:
        "200": { description: Agent decision }
        "400": { description: Invalid request }
        "402": { description: Payment required }
...